backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.19
pydantic==2.10.3
pydantic-settings==2.7.0
pyjwt[crypto]==2.10.1
httpx==0.28.1
openai==1.58.1
pinecone==5.4.2
google-cloud-storage==2.19.0
google-cloud-firestore==2.19.0
pypdf==5.1.0
